// autogpt_platform/frontend/src/app/api/__generated__/models/storeSubmissionRequest.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { StoreSubmissionRequestVideoUrl } from "./storeSubmissionRequestVideoUrl";
import type { StoreSubmissionRequestChangesSummary } from "./storeSubmissionRequestChangesSummary";
import type { StoreSubmissionRequestRecommendedScheduleCron } from "./storeSubmissionRequestRecommendedScheduleCron";

export interface StoreSubmissionRequest {
  agent_id: string;
  agent_version: number;
  slug: string;
  name: string;
  sub_heading: string;
  video_url?: StoreSubmissionRequestVideoUrl;
  image_urls?: string[];
  description?: string;
  categories?: string[];
  changes_summary?: StoreSubmissionRequestChangesSummary;
  recommended_schedule_cron?: StoreSubmissionRequestRecommendedScheduleCron;
}